+++ .everylintrc.js
@@ -1,13 +1,13 @@
 module.exports = {
   javascript: {
     rules: {
-      'semi': 2
-    }
+      semi: 2,
+    },
   },
   markdown: {
     // ...
   },
   style: {
     // ...
-  }
-}
+  },
+};