+++ src/linters/html/index.js
@@ -4,7 +4,7 @@ export default class HTMLLinter {
   static type = 'html';
 
   constructor(/* config */) {
-    this.linter = (content) => HTMLHint.verify(content);
+    this.linter = content => HTMLHint.verify(content);
   }
 
   matchType(file) {
@@ -15,8 +15,14 @@ export default class HTMLLinter {
   async lint(file) {
     const report = this.linter(file.toString());
 
-    report.forEach((problem) => {
-      const { message, line, col: column, rule: { id: ruleId }, type } = problem;
+    report.forEach(problem => {
+      const {
+        message,
+        line,
+        col: column,
+        rule: { id: ruleId },
+        type,
+      } = problem;
       // { type: 'error',
       //   message: 'Doctype must be declared first.',
       //   raw: 'asdf\n',
@@ -27,7 +33,6 @@ export default class HTMLLinter {
       //    { id: 'doctype-first',
       //      description: 'Doctype must be declared first.',
       //      link: 'https://github.com/yaniswang/HTMLHint/wiki/doctype-first' } }
-
 
       if (type === 'error') {
         file.error(message, { line, column }, ruleId);