+++ src/linters/index.js
@@ -3,9 +3,4 @@ import javascript from './javascript';
 import markdown from './markdown';
 import style from './style';
 
-export default [
-  html,
-  javascript,
-  markdown,
-  style,
-];
+export default [html, javascript, markdown, style];