+++ src/linters/style/config.js
@@ -1,11 +1,11 @@
-let config = {}
+let config = {};
 
 /* "stylelint-config-standard" is optional, so, check it */
 try {
-  require.resolve('stylelint-config-standard')
+  require.resolve('stylelint-config-standard');
   config = {
-    extends: 'stylelint-config-standard'
-  }
+    extends: 'stylelint-config-standard',
+  };
 } catch (error) {
   console.error(error);
 }